+++ include/kern/rendering/camera.hpp
@@ -10,38 +10,38 @@ namespace kern::rendering
 {
     class Camera
     {
-        glm::dvec3 position_;
-        glm::dvec3 forward_;
-        glm::dvec3 backward_{};
-        double fov_;
+        glm::vec3 position_;
+        glm::vec3 forward_;
+        glm::vec3 backward_{};
+        float fov_;
 
         // Computed values
-        glm::dvec3 right_{};
-        glm::dvec3 left_{};
+        glm::vec3 right_{};
+        glm::vec3 left_{};
 
-        const glm::dvec3 up_ = glm::dvec3(0.0, 1.0, 0.0);
-        const glm::dvec3 down_ = -up_;
+        const glm::vec3 up_ = glm::vec3(0, 1, 0);
+        const glm::vec3 down_ = -up_;
 
     public:
-        explicit Camera(const glm::dvec3& position, const glm::dvec3& forward, double fov);
+        explicit Camera(const glm::vec3& position, const glm::vec3& forward, float fov);
 
         void recompute_direction_vectors();
 
-        void move(const glm::dvec3& direction, double amount);
-        void move_forward(double amount);
-        void move_backward(double amount);
-        void move_right(double amount);
-        void move_left(double amount);
-        void move_up(double amount);
-        void move_down(double amount);
+        void move(const glm::vec3& direction, float amount);
+        void move_forward(float amount);
+        void move_backward(float amount);
+        void move_right(float amount);
+        void move_left(float amount);
+        void move_up(float amount);
+        void move_down(float amount);
 
-        void set_position(const glm::dvec3& position);
-        void set_direction(const glm::dvec3& direction);
+        void set_position(const glm::vec3& position);
+        void set_direction(const glm::vec3& direction);
 
-        [[nodiscard]] glm::dvec3 get_position() const;
-        [[nodiscard]] glm::dvec3 get_forward() const;
-        [[nodiscard]] glm::dvec3 get_right() const;
-        [[nodiscard]] glm::dvec3 get_left() const;
-        [[nodiscard]] glm::dvec3 get_up() const;
+        [[nodiscard]] glm::vec3 get_position() const;
+        [[nodiscard]] glm::vec3 get_forward() const;
+        [[nodiscard]] glm::vec3 get_right() const;
+        [[nodiscard]] glm::vec3 get_left() const;
+        [[nodiscard]] glm::vec3 get_up() const;
     };
 }
